refactor: controller 및 service 리팩토링
