+++ clog-admin/src/main/resources/application.yml
@@ -11,19 +11,10 @@ spring:
       filter:
         enabled: true
 
-  devtools:
-    livereload:
-      enabled: true
-
   thymeleaf:
     prefix: classpath:templates/
     suffix: .html
     cache: false
-
-jwt:
-  secret: ${JWT_SECRET}
-  access-token-expiration-millis: 3600000
-  refresh-token-expiration-millis: 604800000
 
 ncp:
   storage:
